split truncated_normal, test=kunlun
