+++ paddle/phi/kernels/xpu/truncated_gaussian_random_kernel.cc
@@ -21,6 +21,7 @@ limitations under the License. */
 #include "paddle/phi/backends/cpu/cpu_context.h"
 #include "paddle/phi/backends/xpu/xpu_context.h"
 #include "paddle/phi/core/kernel_registry.h"
+#include "paddle/phi/kernels/funcs/truncated_normal.h"
 
 namespace phi {
